@@ -1,33 +1,31 @@
 {
   "name": "sam-react-node-boilerplate-frontend",
-  "version": "0.0.1",
+  "version": "0.1.0",
   "author": "Craig Barratt",
   "description": "SAM React Node Boilerplate - Frontend",
   "license": "MIT",
   "scripts": {
     "start": "PORT=5555 react-scripts start",
-    "build": "NODE_ENV=production react-scripts test --coverage --watchAll=false && react-scripts build",
+    "build": "NODE_ENV=production react-scripts build",
     "test": "react-scripts test",
     "test:ci": "yarn test --coverage --watchAll=false",
     "lint": "eslint 'src/**/*.{js,jsx}'",
-    "lint:fix": "yarn lint --fix",
-    "generate:component": "npx generate-react-cli component"
+    "lint:fix": "yarn lint --fix"
   },
   "dependencies": {
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "react-redux": "^7.2.2",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "redux-persist": "^6.0.0",
-    "web-vitals": "^1.0.1"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.4.2",
+    "react-scripts": "5.0.1",
+    "web-vitals": "^2.1.4"
   },
   "devDependencies": {
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
+    "eslint": "^8.31.0",
     "eslint-config-prettier": "6.15.0",
-    "eslint-plugin-jest": "24.1.3",
+    "eslint-plugin-jest": "^27.2.1",
     "eslint-plugin-jest-formatting": "^2.0.0",
     "eslint-plugin-jsx-a11y": "^6.4.1",
     "eslint-plugin-prettier": "3.1.4",
