@@ -1,11 +1,9 @@
-import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
+import { Route, Routes } from 'react-router-dom';
 import { Home } from '../../pages/Home/Home';
 import './App.css';
 
 export const App = () => (
-  <Router>
-    <Switch>
-      <Route path="/" exact component={Home} />
-    </Switch>
-  </Router>
+  <Routes>
+    <Route path="/" element={<Home />} />
+  </Routes>
 );
